* master: (162 commits)
  Atualização do Documento de Acompanhamento da OS 4782.
  Atualização do Documento de Acompanhamento da OS 4777.
  Atualização do Documento de Acompanhamento da OS 4776.
  Alteração do Documento de Acompanhamento da OS 4721.
  Publicação de nova versão do casos de uso ARRUC0210 - Gerar DARE-e e ARRUC0240 - Processar Barra DARE, do documento de modelo de casos de usos ARR - Casos de Usos do Processo da Arrecadação - ARR-0200 - DARE-e, e também do ARR - Dicionário de Dados e ARR - Modelo de Dados - Diagrama, para mudança do nome RENAVAN para RENAVAM, inclusão do campo Data da Criação da Conta, ajuste na regra de negócio ARRRN0208 – Calculando Valor TSE a Pagar, e criação de nova regra de negócio ARRRN0212 – Procedimento para Inclusão de Itens de DARE que apresentam Imposto + TSE. Atualização do Documento de Acompanhamento da OS 4797.
  Atualização do documento de Acompanhamento da OS 4797.
  Atualização do Boletim de Acompanhamento e Relatório Mensal de Pagamentos das OS - Ref. 2ª Quinzena de Out/2016.
  Atualização do documento de Acompanhamento da OS 4813.
  Atualização do documento de Acompanhamento da OS 4810.
  Atualização do documento de Acompanhamento da OS 4809.
  Atualização do documento de Acompanhamento da OS 4808.
  Atualização do documento de Acompanhamento da OS 4797.
  Atualização do documento de Acompanhamento da OS 4782.
  Atualização do documento de Acompanhamento da OS 4777.
  Atualização do documento de Acompanhamento da OS 4776.
  Atualização do documento de Acompanhamento da OS 4757.
  Atualização do documento de Acompanhamento da OS 4721.
  Atualização do Controle de Versão
  Entrega dos bugs Não Graves da OS4782.
  Publicação da Planilha de Contagem de Pontos de Função, Parecer Técnico sobre a variação da contagem de pontos de função e atualizado do cronograma da OS, por conta de ajustes nos pontos de função, com o acréscimo de 17 PF por alteração de casos de uso.
  ...

+++ 02_FONTES/core/arrecadacao-core/src/main/java/br/gov/to/sefaz/arr/parametros/business/service/validator/PedidoAreasQtdDiasByTipoValidator.java
@@ -1,9 +1,9 @@
 package br.gov.to.sefaz.arr.parametros.business.service.validator;
 
-import br.gov.to.sefaz.arr.parametros.persistence.entity.PedidoAreas;
-import br.gov.to.sefaz.arr.parametros.persistence.entity.PedidoTipos;
-import br.gov.to.sefaz.arr.parametros.persistence.repository.PedidoAreasRepository;
-import br.gov.to.sefaz.arr.parametros.persistence.repository.PedidoTiposRepository;
+import br.gov.to.sefaz.arr.persistence.entity.PedidoAreas;
+import br.gov.to.sefaz.arr.persistence.entity.PedidoTipos;
+import br.gov.to.sefaz.arr.persistence.repository.PedidoAreasRepository;
+import br.gov.to.sefaz.arr.persistence.repository.PedidoTiposRepository;
 import br.gov.to.sefaz.business.service.validation.ServiceValidator;
 import br.gov.to.sefaz.business.service.validation.ValidationContext;
 import br.gov.to.sefaz.business.service.validation.violation.CustomViolation;
@@ -48,7 +48,7 @@ public class PedidoAreasQtdDiasByTipoValidator implements ServiceValidator<Pedid
 
         PedidoTipos tipo = tiposRepository.findOne(target.getIdTipoPedido());
 
-        Integer totalDias;
+        Long totalDias;
         if (Objects.isNull(target.getIdPedidoArea())) {
             totalDias = repository.getTotalQtdDiasAnaliseByTipo(target.getIdTipoPedido());
         } else {
